requirements.txt
# core
streamlit~=1.25.0
st-pages~=0.4.3

# app: Omymodels Online
omymodels~=0.13.0

# app: Kubernetes Secret
python-dotenv~=1.0.0